ask before deleting a participation

+++ src/app/participation/participation.component.ts
@@ -1,27 +1,15 @@
 import { Component, OnInit } from '@angular/core';
-import { ValidatorFn, AbstractControl, ValidationErrors } from '@angular/forms';
-import { MAT_DATE_LOCALE } from '@angular/material/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { BSON } from 'realm-web';
 import { Participation, snackBarConfig } from '../common/data';
 import { ParticipationEnd, ParticipationService } from '../service/participation.service';
 import { EndPariticipationDialog as EndParticipationDialog } from './end-participation.component';
-
-export function createIbanValidator(): ValidatorFn {
-  const regexp: RegExp = /^([A-Z]{2})(\d{2})(\d{18})$/gm;
-
-  return (control: AbstractControl): ValidationErrors | null =>
-    regexp.test(control.value) ? null : { iban: true };
-}
+import { DeletePariticipationDialog as RemoveParticipationDialog } from './delete-participation.component';
 
 @Component({
   selector: 'app-participation',
-  templateUrl: './participation.component.html',
-  styleUrls: ['./participation.component.css'],
-  providers: [
-    { provide: MAT_DATE_LOCALE, useValue: 'de-DE' },
-  ]
+  templateUrl: './participation.component.html'
 })
 export class ParticipationComponent implements OnInit {
 
@@ -50,17 +38,31 @@ export class ParticipationComponent implements OnInit {
     });
   }
 
-  onRemoveParticipation(id: BSON.ObjectID): void {
-    this.participationService.removeParticipation(id!)
-      .subscribe({
-        next: result => {
-          this.participationService.refetch();
-          this.snackBar.open(`Teilnahme wurde entfernt`, 'Schließen', snackBarConfig);
-        },
-        error: error => {
-          this.snackBar.open(`Teilnahme mit der ID ${id} konnte nicht entfernt werden: ${error}`, 'Schließen', snackBarConfig);
-        }
-      });
+  onDeleteParticipation(id: BSON.ObjectID): void {
+    var participation = this.participationService.getParticipation(id);
+    if (participation === undefined) {
+      this.snackBar.open(`Für die ID ${id} gibt es keine Teilnahme`, 'Schließen', snackBarConfig);
+      return;
+    }
+
+    const dialogRef = this.dialog
+      .open(RemoveParticipationDialog, { data: participation });
+
+    dialogRef.afterClosed().subscribe((remove: boolean) => {
+      if (remove === undefined || !remove)
+        return;
+
+      this.participationService.removeParticipation(id!)
+        .subscribe({
+          next: result => {
+            this.participationService.refetch();
+            this.snackBar.open(`Teilnahme wurde entfernt`, 'Schließen', snackBarConfig);
+          },
+          error: error => {
+            this.snackBar.open(`Teilnahme mit der ID ${id} konnte nicht entfernt werden: ${error}`, 'Schließen', snackBarConfig);
+          }
+        });
+    });
   }
 
   onEndParticipation(id: BSON.ObjectID): void {
